moved all registry_\*() methods to registry.py from functions.py
